==> 01_preprocess.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("data/accepted_2007_to_2018.csv", nrows=500000)
df.head()



cols = [
    'loan_amnt','int_rate','term','emp_length','home_ownership','annual_inc',
    'purpose','dti','delinq_2yrs','inq_last_6mths','open_acc','pub_rec',
    'revol_bal','revol_util','total_acc','fico_range_low','fico_range_high',
    'loan_status'
]

df = df[cols]
df.head()



df['int_rate'] = df['int_rate'].astype(str).str.rstrip('%').astype(float)
df['term'] = df['term'].astype(str).str.extract('(\d+)').astype(float)

def parse_emp(x):
    if pd.isna(x): 
        return np.nan
    x = str(x)
    if x == '10+ years': return 10
    if x == '< 1 year': return 0.5
    try:
        return float(x.split()[0])
    except:
        return np.nan

df['emp_length'] = df['emp_length'].apply(parse_emp)
df.head()


df = df[df['loan_status'].isin(['Fully Paid', 'Charged Off'])]
df['target'] = df['loan_status'].apply(lambda x: 0 if x == 'Fully Paid' else 1)
df['target'].value_counts()



num_cols = df.select_dtypes(include=['float','int']).columns.tolist()
num_cols.remove('target')

for col in num_cols:
    df[col] = df[col].fillna(df[col].median())

cat_cols = ['purpose','home_ownership']

for col in cat_cols:
    df[col] = df[col].fillna(df[col].mode()[0])


df = pd.get_dummies(df, columns=['purpose','home_ownership'], drop_first=True)
df.head()


from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

X = df.drop(columns=['loan_status', 'target'])
y = df['target']

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)

scaler = StandardScaler()
X_train_s = scaler.fit_transform(X_train)
X_test_s = scaler.transform(X_test)



import joblib
joblib.dump((X_train_s, X_test_s, y_train, y_test, scaler), "preprocessed.joblib")


loan_info = df.loc[y_test.index, ['loan_amnt','int_rate','loan_status']]
loan_info.to_csv("loan_info_test.csv", index=False)
